DTOs renamed, deleted & updated
