+++ OngProject/Core/Models/DTOs/Comments/CommentCreateDTO.cs
@@ -1,6 +1,6 @@
 using System.ComponentModel.DataAnnotations;
 
-namespace OngProject.Core.Models.DTOs
+namespace OngProject.Core.Models.DTOs.Comments
 {
     public class CommentCreateDTO
     {
