+++ OngProject/Core/Models/DTOs/Comments/CommentDTO.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace OngProject.Core.Models.DTOs
+namespace OngProject.Core.Models.DTOs.Comments
 {
     public class CommentDTO
     {
