+++ OngProject/Core/Models/DTOs/News/NewsGetDTO.cs
@@ -1,0 +1,16 @@
+using System;
+
+namespace OngProject.Core.Models.DTOs.News
+{
+    public class NewsGetDTO
+    {
+        public int Id { get; set; }
+        public string Name { get; set; }
+        public string Content { get; set; }
+        public string Image { get; set; }
+        public DateTime CreationDate { get; set; }
+        public DateTime ExpirationDate { get; set; }
+    }
+
+}
+
