+++ OngProject/Core/Models/DTOs/Slides/SlideDetailsDTO.cs
@@ -1,7 +1,7 @@
 using System;
 
 
-namespace OngProject.Core.Models.DTOs
+namespace OngProject.Core.Models.DTOs.Slides
 {
     public class SlideDetailsDTO
     {
