--- OngProject/Core/Models/Users/UserModel.cs
@@ -1,21 +1,0 @@
-using System;
-
-namespace OngProject.Core.Models.Users
-{
-    public class UserModel
-    {
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
-        public string Email { get; set; }
-        public string Password { get; set; }
-        public string Photo { get; set; }
-        public byte[] PasswordHash {get;set;}
-        public byte[] PasswordSalt {get;set;}
-        public DateTime CreatedAt { get; set; }
-        public DateTime? UpdatedAt { get; set; }
-        public bool IsDeleted { set; get; }
-    }
-}
-
-
-
